♻️ Refactor PlayerCharacter types
